% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{average_consumption}
\alias{average_consumption}
\title{Average of energy consumed per activation period}
\usage{
average_consumption(steps_record)
}
\arguments{
\item{steps_record}{A dataframe of events record}
}
\value{
A value. Energy (Wh)
}
\description{
Average of energy consumed per activation period
}
