% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{average_duration}
\alias{average_duration}
\title{Average duration (hours) of device activation period}
\usage{
average_duration(steps_record)
}
\arguments{
\item{steps_record}{A dataframe of events record}
}
\value{
A value. Duration (hours).
}
\description{
Average duration (hours) of device activation period
}
