% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{average_start_time}
\alias{average_start_time}
\title{Average session starting time}
\usage{
average_start_time(steps_record)
}
\arguments{
\item{steps_record}{A dataframe of events record}
}
\value{
A value
}
\description{
Average session starting time
}
