% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{device_consumption_plot}
\alias{device_consumption_plot}
\title{Dygrpahs plot object of device cosumption over total consumption}
\usage{
device_consumption_plot(historical_consumption, device_power)
}
\arguments{
\item{historical_consumption}{A dataframe with datetime and consumption columns}

\item{device_power}{A number corresponding to device constant power (W)}

\item{user_name}{Prosumer name}
}
\value{
Dygraphs plot object
}
\description{
Dygrpahs plot object of device cosumption over total consumption
}
