% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esteps.R
\name{get_average_consumption}
\alias{get_average_consumption}
\title{Average of energy consumed per activation period}
\usage{
get_average_consumption(steps_record)
}
\arguments{
\item{steps_record}{A tibble of events record returned by \code{get_steps_record}}
}
\value{
A value. Energy (Wh)
}
\description{
Average of energy consumed per activation period
}
