% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esteps.R
\name{get_average_duration}
\alias{get_average_duration}
\title{Average duration (minutes) of device activation period}
\usage{
get_average_duration(steps_record)
}
\arguments{
\item{steps_record}{A tibble of events record returned by \code{get_steps_record}}
}
\value{
Mean of duration values (minutes).
}
\description{
Average duration (minutes) of device activation period
}
