% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esteps.R
\name{get_average_start_time}
\alias{get_average_start_time}
\title{Average session starting time}
\usage{
get_average_start_time(steps_record)
}
\arguments{
\item{steps_record}{A tibble of events record returned by \code{get_steps_record}}
}
\value{
A value
}
\description{
Average session starting time
}
