% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esteps.R
\name{get_number_sessions}
\alias{get_number_sessions}
\title{Number of sessions detected}
\usage{
get_number_sessions(steps_record)
}
\arguments{
\item{steps_record}{A tibble of events record returned by \code{get_steps_record}}
}
\value{
A value
}
\description{
Number of sessions detected
}
