% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esteps.R
\name{get_steps}
\alias{get_steps}
\title{Detect steps of constant power}
\usage{
get_steps(historical_consumption, device_power)
}
\arguments{
\item{historical_consumption}{A tibble with \code{datetime} and \code{power} columns}

\item{device_power}{A number corresponding to device constant power}
}
\value{
A tibble with the activation and deactivation time of all detected steps
}
\description{
Detect steps of constant power
}
\keyword{internal}
