% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esteps.R
\name{get_steps_record}
\alias{get_steps_record}
\title{Summarize all events detected}
\usage{
get_steps_record(historical_consumption, device_power, at_least)
}
\arguments{
\item{historical_consumption}{A tibble with \code{datetime} and \code{power} columns}

\item{device_power}{A number corresponding to device constant power}

\item{at_least}{Minimum of minutes to consider the step}
}
\value{
A tibble of events record, composed by:
activation datetime
deactivation datetime
total consumption within this period
device consumption whithin this period
}
\description{
Summarize all events detected
}
