% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{historical_consumption}
\alias{historical_consumption}
\title{Historical consumption example}
\format{
\subsection{\code{historical_consumption}}{

A data frame with 11,371 rows and 2 columns:
\describe{
\item{datetime}{Datetime sequence with Amsterdam timezone}
\item{power}{Power profile in W}
...
}
}
}
\source{
Eastern Docklands Prosumer Community
}
\usage{
historical_consumption
}
\description{
An example of power demand from a household with a heat pump
}
\keyword{datasets}
