% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{number_sessions}
\alias{number_sessions}
\title{Number of sessions detected}
\usage{
number_sessions(steps_record)
}
\arguments{
\item{steps_record}{A dataframe of events record}
}
\value{
A value
}
\description{
Number of sessions detected
}
