% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esteps.R
\name{plot_device_consumption}
\alias{plot_device_consumption}
\title{Dygrpahs plot object of device cosumption over total consumption}
\usage{
plot_device_consumption(historical_consumption, device_power, at_least = 60)
}
\arguments{
\item{historical_consumption}{A tibble with \code{datetime} and \code{power} columns}

\item{device_power}{A number corresponding to device constant power}

\item{at_least}{Minimum of minutes to consider the step}
}
\value{
Dygraphs plot object
}
\description{
Dygrpahs plot object of device cosumption over total consumption
}
