% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esteps.R
\name{plot_session_duration_histogram}
\alias{plot_session_duration_histogram}
\title{Histogram of sessions duration}
\usage{
plot_session_duration_histogram(steps_record)
}
\arguments{
\item{steps_record}{Summary from device events summary function}
}
\value{
Histogram plot of steps starting hour
}
\description{
Histogram of sessions duration
}
