% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esteps.R
\name{plot_start_time_histogram}
\alias{plot_start_time_histogram}
\title{Histogram of sessions starting hours}
\usage{
plot_start_time_histogram(steps_record)
}
\arguments{
\item{steps_record}{Summary from device events summary function}
}
\value{
Histogram plot of steps starting hour
}
\description{
Histogram of sessions starting hours
}
