% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{session_duration_histogram}
\alias{session_duration_histogram}
\title{Histogram of sessions duration}
\usage{
session_duration_histogram(user_name, steps_record)
}
\arguments{
\item{user_name}{Prosumer name}

\item{steps_record}{Summary from device events summary function}
}
\value{
Histogram plot of steps starting hour
}
\description{
Histogram of sessions duration
}
