% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{start_time_histogram}
\alias{start_time_histogram}
\title{Histogram of sessions starting hours}
\usage{
start_time_histogram(user_name, steps_record)
}
\arguments{
\item{user_name}{Prosumer name}

\item{steps_record}{Summary from device events summary function}
}
\value{
Histogram plot of steps starting hour
}
\description{
Histogram of sessions starting hours
}
