% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{steps_record}
\alias{steps_record}
\title{Summarize all events detected}
\usage{
steps_record(historical_consumption, device_power)
}
\arguments{
\item{historical_consumption}{A dataframe with datetime and consumption columns}

\item{device_power}{A number corresponding to device constant power}
}
\value{
A dataframe of events record, composed by:
    activation datetime
    deactivation datetime
    total consumption within this period
    device consumption whithin this period
}
\description{
Summarize all events detected
}
