% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{steps_summary}
\alias{steps_summary}
\title{Summary of device sessions indicators}
\usage{
steps_summary(historical_consumption, device_power)
}
\arguments{
\item{historical_consumption}{A dataframe with datetime and consumption columns}

\item{device_power}{A number corresponding to device constant power (W)}
}
\value{
A dataframe with the following indicators:
- Average consumption per session (Wh)
- Average duration of a session (hours)
- Average sessions starting time
- Number of sessions in the time studied
}
\description{
Summary of device sessions indicators
}
