% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esteps.R
\name{summarise_steps}
\alias{summarise_steps}
\title{Summary of device sessions indicators}
\usage{
summarise_steps(historical_consumption, device_power, at_least = 60)
}
\arguments{
\item{historical_consumption}{A tibble with \code{datetime} and \code{power} columns}

\item{device_power}{A number corresponding to device constant power}

\item{at_least}{Minimum of minutes to consider the step}
}
\value{
A tibble with the following indicators:
\itemize{
\item Average consumption per session (Wh)
\item Average duration of a session (hours)
\item Average sessions starting time
\item Number of sessions in the time studied
}
}
\description{
Summary of device sessions indicators
}
